+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## Unreleased
+## 4.0.1 - 2026-09-23
 
 ### Changed
 - Render author-entered node templates in proper Twig sandbox.